requirements.txt
Django==1.10.2
djangorestframework==3.5.1
whitenoise==3.2.2
